refactor: using backstratcher as plural
